+++ payjoin/Cargo.toml
@@ -41,6 +41,7 @@ serde_json = "1.0.108"
 bitcoind = { version = "0.36.0", features = ["0_21_2"] }
 http = "1"
 payjoin-directory = { path = "../payjoin-directory", features = ["_danger-local-https"] }
+payjoin-test-utils = { path = "../payjoin-test-utils" }
 ohttp-relay = "0.0.8"
 once_cell = "1"
 rcgen = { version = "0.11" }
